+++ src/commonMain/kotlin/response/RecentHistory.kt
@@ -6,7 +6,7 @@ import kotlinx.serialization.Serializable
 @JsExport @Serializable class RecentHistory(
     val hq: Boolean,
     val pricePerUnit: Int,
-    val quantity: Byte,
+    val quantity: Int,
     val timestamp: Long,
     val onMannequin: Boolean,
     val buyerName: String,
